requirements.txt
alembic==1.7.7
altair==4.2.0
anyio==3.5.0
apache-airflow==2.3.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
appnope==0.1.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==20.3.0
Babel==2.10.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==0.7.7
cachelib==0.6.0
cachetools==5.0.0
catalogue==2.0.7
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.4
cryptography==37.0.2
cssselect==1.1.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
# en-core-web-sm==3.2.0
entrypoints==0.4
executing==0.8.3
fastapi==0.76.0
fastjsonschema==2.15.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.27
gnewsclient==1.12
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==1.1.0
jedi==0.18.1
jieba3k==0.35.1
Jinja2<3.0 
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keybert==0.5.1
langcodes==3.3.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.8.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
newspaper3k==0.2.8
nltk==3.7
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
preshed==3.0.6
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.11.2
PyJWT==1.7.1
pymango==0.1.1
pymongo==4.1.1
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
rfc3986==1.5.0
rich==12.3.0
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
setproctitle==1.2.3
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-streamlit==1.0.3
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
srsly==2.4.3
stack-data==0.2.0
starlette==0.18.0
streamlit==1.8.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
thinc==8.0.15
threadpoolctl==3.1.0
tinycss2==1.1.1
tinysegmenter==0.3
tldextract==3.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
typer==0.4.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
unicodecsv==0.14.1
urllib3==1.26.9
uvicorn==0.17.6
validators==0.18.2
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
matplotlib
streamlit_authenticator
pymongo==4.1.1
requests==2.27.1
flair
setuptools==45
seaborn
streamlit_authenticator
